+++ core/pioreactor/calibrations/__init__.py
@@ -4,18 +4,21 @@ from __future__ import annotations
 import os
 from collections import defaultdict
 from pathlib import Path
-from typing import Callable
+from typing import Any
+from typing import ClassVar
+from typing import Generic
 from typing import Literal
 from typing import overload
-from typing import Type
+from typing import TypeVar
 
 from msgspec import ValidationError
 from msgspec.yaml import decode as yaml_decode
 from msgspec.yaml import encode as yaml_encode
 from pioreactor import structs
-from pioreactor.types import PumpCalibrationDevices
+from pioreactor import types as pt
 from pioreactor.utils import local_persistent_storage
 from pioreactor.whoami import is_testing_env
+
 
 if not is_testing_env():
     CALIBRATION_PATH = Path("/home/pioreactor/.pioreactor/storage/calibrations/")
@@ -23,16 +26,16 @@ else:
     CALIBRATION_PATH = Path(os.environ["DOT_PIOREACTOR"]) / "storage" / "calibrations"
 
 # Lookup table for different calibration protocols
-Device = str
+Device = TypeVar("Device", bound=str)
 ProtocolName = str
 
-calibration_protocols: dict[Device, dict[ProtocolName, Type[CalibrationProtocol]]] = defaultdict(dict)
+calibration_protocols: dict[str, dict[ProtocolName, type[CalibrationProtocol[Any]]]] = defaultdict(dict)
 
 
-class CalibrationProtocol:
-    protocol_name: ProtocolName
-    target_device: Device | list[Device]
-    description = ""
+class CalibrationProtocol(Generic[Device]):
+    protocol_name: ClassVar[ProtocolName]
+    target_device: ClassVar[str | list[str]]
+    description: ClassVar[str] = ""
 
     def __init_subclass__(cls, **kwargs):
         super().__init_subclass__(**kwargs)
@@ -44,48 +47,52 @@ class CalibrationProtocol:
         else:
             raise ValueError("target_device must be a string or a list of strings")
 
-    def run(self, *args, **kwargs) -> structs.CalibrationBase:
+    def run(self, target_device: Device) -> structs.CalibrationBase | list[structs.CalibrationBase]:
         raise NotImplementedError("Subclasses must implement this method.")
 
 
-class SingleVialODProtocol(CalibrationProtocol):
-    target_device = "od"
+class SingleVialODProtocol(CalibrationProtocol[pt.ODCalibrationDevices]):
+    target_device = pt.OD_DEVICES
     protocol_name = "single_vial"
     description = "Calibrate OD using a single vial"
 
-    def run(self, *args, **kwargs) -> structs.OD600Calibration:
+    def run(self, target_device: pt.ODCalibrationDevices, **kwargs) -> structs.OD600Calibration:
         from pioreactor.calibrations.od_calibration_single_vial import run_od_calibration
 
-        return run_od_calibration()
+        return run_od_calibration(target_device)
 
 
-class StandardsODProtocol(CalibrationProtocol):
-    target_device = "od"
+class StandardsODProtocol(CalibrationProtocol[pt.ODCalibrationDevices]):
+    target_device = pt.OD_DEVICES
     protocol_name = "standards"
     description = "Calibrate OD using standards. Requires multiple vials"
 
-    def run(self, *args, **kwargs) -> structs.OD600Calibration:
+    def run(  # type: ignore
+        self, target_device: pt.ODCalibrationDevices, *args, **kwargs
+    ) -> structs.OD600Calibration | list[structs.OD600Calibration]:
         from pioreactor.calibrations.od_calibration_using_standards import run_od_calibration
 
-        return run_od_calibration()
+        return run_od_calibration(target_device)
 
 
-class DurationBasedPumpProtocol(CalibrationProtocol):
-    target_device = ["media_pump", "alt_media_pump", "waste_pump"]
+class DurationBasedPumpProtocol(CalibrationProtocol[pt.PumpCalibrationDevices]):
+    target_device = pt.PUMP_DEVICES
     protocol_name = "duration_based"
 
-    def run(self, target_device: str, **kwargs) -> structs.SimplePeristalticPumpCalibration:
+    def run(
+        self, target_device: pt.PumpCalibrationDevices, **kwargs
+    ) -> structs.SimplePeristalticPumpCalibration:
         from pioreactor.calibrations.pump_calibration import run_pump_calibration
 
         return run_pump_calibration(target_device)
 
 
-class DCBasedStirringProtocol(CalibrationProtocol):
+class DCBasedStirringProtocol(CalibrationProtocol[Literal["stirring"]]):
     target_device = "stirring"
     protocol_name = "dc_based"
 
     def run(
-        self, target_device: str, min_dc: str | None = None, max_dc: str | None = None
+        self, target_device: Literal["stirring"], min_dc: str | None = None, max_dc: str | None = None
     ) -> structs.SimpleStirringCalibration:
         from pioreactor.calibrations.stirring_calibration import run_stirring_calibration
 
@@ -95,13 +102,13 @@ class DCBasedStirringProtocol(CalibrationProtocol):
 
 
 @overload
-def load_active_calibration(device: Literal["od"]) -> structs.ODCalibration | None:
+def load_active_calibration(device: pt.ODCalibrationDevices) -> structs.ODCalibration | None:
     pass
 
 
 @overload
 def load_active_calibration(
-    device: PumpCalibrationDevices,
+    device: pt.PumpCalibrationDevices,
 ) -> structs.SimplePeristalticPumpCalibration | None:
     pass